Move `transformers.py` to `huggingface/transformers.py`
